requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
cssselect==1.3.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dnspython==2.7.0
fake-useragent==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
geopandas==1.0.1
h11==0.16.0
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
license-expression==30.4.1
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
nbformat==5.10.4
numpy==2.2.6
outcome==1.3.0.post0
packageurl-python==0.16.0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
Protego==0.4.0
psutil==7.0.0
py-serializable==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pymongo==4.13.0
pyogrio==0.11.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyproj==3.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
queuelib==1.8.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rpds-py==0.25.1
Scrapy==2.13.1
scrapyd==1.5.0
selenium==4.33.0
service-identity==24.2.0
session_info==1.0.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
stdlib-list==0.11.1
tldextract==5.3.0
toml==0.10.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
